Use NIST SRD 8 to cover all elements

+++ openmc/material.py
@@ -412,8 +412,6 @@ class Material(IDManagerMixin):
 
         return combined
 
-
-
     def get_photon_contact_dose_rate(
         self,
         dose_quantity: str = "absorbed-air",
@@ -589,12 +587,12 @@ class Material(IDManagerMixin):
                 integrand = (response_f(e_vals) / mu_vals) * p_vals
 
             if isinstance(photon_source_per_atom, Discrete):
-                cdr_nuc = float(np.sum(integrand))
+                cdr_nuc = np.sum(integrand)
             elif isinstance(photon_source_per_atom, Tabular):
-                cdr_nuc = float(np.trapezoid(integrand, e_vals))
+                cdr_nuc = np.trapezoid(integrand, e_vals)
 
             # Compute air-absorbed dose [Gy/h] or effective dose [Sv/h]
-            cdr[nuc] = cdr_nuc * nuc_atoms_per_bcm * multiplier
+            cdr[nuc] = float(cdr_nuc * nuc_atoms_per_bcm * multiplier)
 
         return cdr if by_nuclide else sum(cdr.values())
